projects/booking_CRM_project/dist_app/requirements.txt
Django>=5.2,<6
djangorestframework>=3.15,<3.16
drf-spectacular>=0.27
django-cors-headers>=4.3
django-filter>=24.2
python-dotenv>=1.0
dj-database-url>=2.0
requests>=2.31
django-csp>=3.8
sentry-sdk>=2.17
python-json-logger>=2.0
celery>=5.4
redis>=5.0
djangorestframework-simplejwt>=5.3
icalendar>=5.0
stripe>=11.3

Pillow>=10.0
 